requirements.txt
openai
faiss-cpu
langchain
python-dotenv
pypdf2
langchain_openai
sentence-transformers
langchain_community
streamlit